requirements.txt
alabaster==1.0.0
babel==2.16.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
docutils==0.21.2
idna==3.10
imagesize==1.4.1
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.2.0
packaging==24.2
Pygments==2.18.0
pyserial==3.5
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
serpent==1.41
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
urllib3==2.2.3
